+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - ready for github release
 
+- Chain getting balances with new header
+
 # v0.6.1 (2019.03.12)
 
 ### Fixes